% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_dir_renv}
\alias{.env_dir_renv}
\title{Internal environment variable: \code{{renv}} dir}
\usage{
.env_dir_renv()
}
\value{

}
\description{
Single-source-of-truth encapsulation of internal settings for
\code{{renv}} directory.
}
\examples{
.env_dir_renv()
}
