% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_dir_renv_cache}
\alias{.env_dir_renv_cache}
\title{Internal environment variable: \code{{renv}} dir: cache}
\usage{
.env_dir_renv_cache()
}
\value{

}
\description{
Single-source-of-truth encapsulation of internal settings for
\code{{renv}} dir for ache
}
