% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_dir_renv_cache_docker}
\alias{.env_dir_renv_cache_docker}
\title{Internal environment variable: \code{{renv}} dir: cache_docker}
\usage{
.env_dir_renv_cache_docker()
}
\value{

}
\description{
Single-source-of-truth encapsulation of internal settings for
\code{{renv}} dir for docker cache
}
\examples{
.env_dir_renv_cache_docker()
}
