% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_dir_renv_local}
\alias{.env_dir_renv_local}
\title{Internal environment variable: \code{{renv}} dir: local}
\usage{
.env_dir_renv_local()
}
\value{

}
\description{
Single-source-of-truth encapsulation of internal settings for
\code{{renv}} directory where tar.gz of the current package is persisted
}
