% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_renv_activate}
\alias{.env_renv_activate}
\title{Internal environment variable: renv/activate}
\usage{
.env_renv_activate()
}
\value{

}
\description{
Exists for single-source-of-truth encapsulation internal settings for
\code{{renv}} activation script
}
\examples{
.env_renv_activate()
}
