% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_renv_lockfile}
\alias{.env_renv_lockfile}
\title{Internal environment variable: renv lockfile}
\usage{
.env_renv_lockfile()
}
\value{

}
\description{
Exists for single-source-of-truth encapsulation of internal settings for
\code{{renv}} lockfile
}
