% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_renv_manifest}
\alias{.env_renv_manifest}
\title{Internal environment variable: renv manifest}
\usage{
.env_renv_manifest()
}
\value{

}
\description{
Exists for single-source-of-truth encapsulation of internal settings for
\code{{renv}} manifest
}
