% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.env_renv_manifest_cached}
\alias{.env_renv_manifest_cached}
\title{Internal environment variable: cached renv manifest}
\usage{
.env_renv_manifest_cached()
}
\value{

}
\description{
Exists for single-source-of-truth encapsulation of internal settings for
\code{{renv}} manifest
}
