% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_dcm_name}
\alias{env_dcm_name}
\title{Environment variable: dependency manager name}
\usage{
env_dcm_name()
}
\value{
[\link{character}] Package name with suffix \verb{_deps}
}
\description{
Environment variable: dependency manager name
}
\examples{
env_dcm_name()
}
