% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_dependency_manager_name}
\alias{env_dependency_manager_name}
\title{Environment variable value: dependency manager name}
\usage{
env_dependency_manager_name()
}
\value{

}
\description{
Environment variable value: dependency manager name
}
\examples{
env_dependency_manager_name()
}
