% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_package_maintainer}
\alias{env_package_maintainer}
\title{Environment variable: package maintainer}
\usage{
env_package_maintainer()
}
\value{

}
\description{
Environment variable: package maintainer
}
\examples{
env_package_maintainer()
}
