% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_package_port}
\alias{env_package_port}
\title{Environment variable: package port}
\usage{
env_package_port(port = 8000)
}
\arguments{
\item{port}{[\link{integer}] Port. Defaults to \code{8000}}
}
\value{
[\link{integer}] Port
}
\description{
Environment variable: package port
}
\examples{
env_package_port()
}
