% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_package_version}
\alias{env_package_version}
\title{Environment variable value: package version}
\usage{
env_package_version()
}
\value{

}
\description{
Environment variable value: package version
}
\examples{
env_package_version()
}
