% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_r_version}
\alias{env_r_version}
\title{Environment variable: r version}
\usage{
env_r_version()
}
\value{
[\link{character}] R version (\verb{\{major\}.\{minor\}.\{patch\}})
}
\description{
Relevant for cached build argument/environment variable for Docker.
}
\examples{
env_r_version()
}
