% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_renv_version}
\alias{env_renv_version}
\title{Environment variable value: renv version}
\usage{
env_renv_version()
}
\value{

}
\description{
Environment variable value: renv version
}
\examples{
env_renv_version()
}
