% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_add_lockfile_record}
\alias{renv_add_lockfile_record}
\title{Add \code{{renv}} lockfile record for current package}
\usage{
renv_add_lockfile_record()
}
\value{

}
\description{
Add \code{{renv}} lockfile record for current package
}
\examples{
renv_add_lockfile_record()
}
