% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_copy_description}
\alias{renv_copy_description}
\title{Ensure directory exists and is empty}
\usage{
renv_copy_description()
}
\value{

}
\description{
Ensure directory exists and is empty
}
\examples{
renv_ensure_local_dir()
}
