% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_digest_cached_lockfile}
\alias{renv_digest_cached_lockfile}
\title{Digest cached \code{{renv}} lockfile}
\usage{
renv_digest_cached_lockfile()
}
\value{

}
\description{
Digest cached \code{{renv}} lockfile
}
