% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_digest_primary_manifest}
\alias{renv_digest_primary_manifest}
\title{Get hash of \code{{renv}} manifest}
\usage{
renv_digest_primary_manifest()
}
\value{

}
\description{
Get hash of \code{{renv}} manifest
}
