% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_ensure_cache_dir}
\alias{renv_ensure_cache_dir}
\title{Ensure directory: renv/cache}
\usage{
renv_ensure_cache_dir()
}
\value{

}
\description{
Ensure directory exists.
}
\examples{
renv_ensure_cache_docker_dir()
}
