% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_ensure_local_dir}
\alias{renv_ensure_local_dir}
\title{Ensure directory: renv/local}
\usage{
renv_ensure_local_dir()
}
\value{

}
\description{
Ensure directory exists and is empty
}
\examples{
renv_ensure_local_dir()
}
