% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_remove_dev_package}
\alias{renv_remove_dev_package}
\title{Remove dev package from \code{{renv}} library}
\usage{
renv_remove_dev_package()
}
\value{

}
\description{
Remove dev package from \code{{renv}} library
}
\examples{
renv_remove_dev_package()
}
