% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_set_env_var_cache}
\alias{renv_set_env_var_cache}
\title{Set environment variable for \code{{renv}} cache}
\usage{
renv_set_env_var_cache()
}
\value{

}
\description{
Set environment variable for \code{{renv}} cache
}
\examples{
renv_set_env_var_cache()
}
