% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_template}
\alias{use_template}
\title{Use template}
\usage{
use_template(
  template,
  save_as = template,
  data = list(),
  ignore = FALSE,
  open = FALSE,
  package = "usethis"
)
}
\arguments{
\item{package}{}
}
\value{

}
\description{
Motivated by \link[usethis:use_template]{usethis::use_template}, but needed to be liberated from
always using \link[usethis:proj_utils]{usethis::proj_path}.
}
