% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_template_build}
\alias{use_template_build}
\title{Use build template}
\usage{
use_template_build(dir = here::here(), open = rlang::is_interactive())
}
\arguments{
\item{dir}{[\link{logical}] Path to directory where file should be created}

\item{open}{[\link{logical}]
\itemize{
\item \code{TRUE}: create and open file
\item \code{FALSE}: create but do not open file
}}
}
\value{
[\link{character}] Path to file that was created based on template
}
\description{
Use build template
}
