% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_template_docker_run}
\alias{use_template_docker_run}
\title{Use docker run template}
\usage{
use_template_docker_run(dir = here::here(), open = rlang::is_interactive())
}
\arguments{
\item{dir}{[\link{character}] Path to directory where file should be created}

\item{open}{[\link{logical}]
\itemize{
\item \code{TRUE}: create and open file
\item \code{FALSE}: create but do not open file
}}
}
\value{
[\link{character}] Path to file that was created based on template
}
\description{
Use docker run template
}
\examples{
use_template_docker_run(dir = tempdir())
}
