% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_template_dockerfile}
\alias{use_template_dockerfile}
\title{Use Dockerfile template}
\usage{
use_template_dockerfile(
  dir = here::here(),
  r_script = "main.R",
  open = rlang::is_interactive()
)
}
\arguments{
\item{dir}{[\link{character}] Path to directory where file should be created}

\item{r_script}{[\link{character}] File name of script within \verb{inst/}}

\item{open}{[\link{logical}]
\itemize{
\item \code{TRUE}: create and open file
\item \code{FALSE}: create but do not open file
}}
}
\value{
[\link{character}] Path to file that was created based on template
}
\description{
Use Dockerfile template
}
\examples{
use_template_dockerfile(dir = tempdir())
}
