% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{write_env_var}
\alias{write_env_var}
\title{Write environment variable to file}
\usage{
write_env_var(value, dir = here::here(), file_name = character())
}
\arguments{
\item{value}{[\link{character}] Value of the env var.}

\item{dir}{[\link{character}] Path to directory that env var files should be
written to. Defaults to root directory of your package/project (via
\code{\link[here:here]{here::here()}})}

\item{file_name}{[\link{character}] Optional file name. If none is provided then
\verb{.docker_env_\{function_name\}} is use with \code{function_name} being the name of
the function that called this function (e.g. \link{env_package_name}).}
}
\value{
[\link{character}] File path
}
\description{
Write environment variable to file
}
