% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_env_vars}
\alias{write_env_vars}
\title{Write environment variables to file}
\usage{
write_env_vars(dir = here::here())
}
\arguments{
\item{dir}{[\link{character}] Path to directory that env var files should be
written to. Defaults to root directory of your package/project (via
\code{\link[here:here]{here::here()}})}
}
\description{
Write environment variables to file
}
\examples{
write_env_vars(dir = tempdir())
}
