% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_digest_primary_lockfile}
\alias{renv_digest_primary_lockfile}
\title{Get hash of \code{{renv}} lockfile}
\usage{
renv_digest_primary_lockfile(path = .env_renv_lockfile())
}
\arguments{
\item{path}{[\link{character}] Path to \code{renv.lock}}
}
\value{

}
\description{
Get hash of \code{{renv}} lockfile
}
\examples{
renv_digest_primary_lockfile()
}
