% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_manage_cached_lockfile}
\alias{renv_manage_cached_lockfile}
\title{Manage cached \code{{renv}} lockfile}
\usage{
renv_manage_cached_lockfile(
  key_primary = renv_digest_primary_lockfile(),
  key_cached = renv_digest_cached_lockfile(),
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\value{
Path to ached \code{{renv}} lockfile
}
\description{
Manage cached \code{{renv}} lockfile
}
\examples{
renv_manage_cached_lockfile()
}
