% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{write_env_vars}
\alias{write_env_vars}
\title{Write environment variables to file}
\usage{
write_env_vars(dir = here::here(), file_name = ".env")
}
\arguments{
\item{dir}{[\link{character}] Path to directory that env var files should be
written to. Defaults to root directory of your package/project (via
\code{\link[here:here]{here::here()}})}
}
\value{
[\link{list}] List of path names
}
\description{
Write environment variables to file
}
\examples{
write_env_vars(dir = tempdir())
}
